feat: 스트릭 클릭 시 피드 모달 뜨는 기능 구현
